requirements.txt
python-dotenv==1.0.1
google-search-results==2.4.2
langchain-community==0.3.8
langchain-google-genai==2.0.5
langchain==0.3.8